// src/compiler/compilerVersions.generated.ts
// deno-fmt-ignore-file
/* Automatically maintained from package.json. Do not edit! */

export type CompilerPackageNames = "typescript-5.5.3" | "typescript-5.4.5" | "typescript-5.3.3" | "typescript-5.2.2" | "typescript-5.1.6" | "typescript-5.0.4" | "typescript-4.9.5" | "typescript-4.8.4" | "typescript-4.7.4" | "typescript-next";
export type CompilerVersions = "5.5.3" | "5.4.5" | "5.3.3" | "5.2.2" | "5.1.6" | "5.0.4" | "4.9.5" | "4.8.4" | "4.7.4" | "@next (5.6.0-dev.20240712)";
export const compilerVersionCollection: { version: CompilerVersions; packageName: CompilerPackageNames; }[] = [
        { version: "5.5.3", packageName: "typescript-5.5.3" },
        { version: "5.4.5", packageName: "typescript-5.4.5" },
        { version: "5.3.3", packageName: "typescript-5.3.3" },
        { version: "5.2.2", packageName: "typescript-5.2.2" },
        { version: "5.1.6", packageName: "typescript-5.1.6" },
        { version: "5.0.4", packageName: "typescript-5.0.4" },
        { version: "4.9.5", packageName: "typescript-4.9.5" },
        { version: "4.8.4", packageName: "typescript-4.8.4" },
        { version: "4.7.4", packageName: "typescript-4.7.4" },
        { version: "@next (5.6.0-dev.20240712)", packageName: "typescript-next" }
    ];